Use index types to number polynomials
This is a substantial API change, although it is not a
breaking change as the old API is still supported via
deprecations. The concept is to use three types, `NM`, `OSA`, and `Noll`, to specify the numbering
convention. The constructor of each checks for
validity, so it is impossible to construct invalid index
values. Then, instead of passing in both an `::Int` and
an `index ∈ (:OSA, :Noll)` to lots of functions, the
package functions are simply implemented in terms
of their preferred index type, and other index types
are supported by conversion.

The advantages of this design are:
- index validity is guaranteed and only has to be checked in one place
- indices are "self-documenting" about how they should be interpreted (unlike an `Int`).